// frontend/src/assets/languages/limboole/ls/generated/ast.ts
export interface Expr extends AstNode {
    readonly $type: 'And' | 'Expr' | 'Iff' | 'Implies' | 'Or';
    var: string;
}

export const Expr = 'Expr';

export function isExpr(item: unknown): item is Expr {
    return reflection.isInstance(item, Expr);
}

export interface And extends Expr {
    readonly $type: 'And';
    left: Expr;
    right: Expr;
}

export const And = 'And';

export function isAnd(item: unknown): item is And {
    return reflection.isInstance(item, And);
}

export interface Iff extends Expr {
    readonly $type: 'Iff';
    left: Expr;
    right: Expr;
}

export const Iff = 'Iff';

export function isIff(item: unknown): item is Iff {
    return reflection.isInstance(item, Iff);
}

export interface Implies extends Expr {
    readonly $type: 'Implies';
    left: Expr;
    right: Expr;
}

export const Implies = 'Implies';

export function isImplies(item: unknown): item is Implies {
    return reflection.isInstance(item, Implies);
}

export interface Or extends Expr {
    readonly $type: 'Or';
    left: Expr;
    right: Expr;
}

export const Or = 'Or';

export function isOr(item: unknown): item is Or {
    return reflection.isInstance(item, Or);
}

export type LimbooleAstType = {
    And: And
    Expr: Expr
    Iff: Iff
    Implies: Implies
    Or: Or
}